backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
Pillow==10.0.1
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.0.0